feat: extract V8 coverage primitives to @mizchi/playwright-v8-coverage
Second Layer-1 extraction following the workspace skeleton in #42 and
the @mizchi/server-faults pilot in #43. Splits the V8 precise-coverage
machinery out of chaosbringer so any Playwright project can pick it
up — the design doc roadmap's Tier-1 candidate.

What moves:
- packages/chaosbringer/src/coverage.ts (182 lines) -> packages/playwright-v8-coverage/src/
- packages/chaosbringer/src/coverage.test.ts (179 lines, 17 tests) -> packages/playwright-v8-coverage/src/

Public surface (unchanged module shape, new module name):
- CoverageCollector — wraps Profiler.startPreciseCoverage / takePreciseCoverage
- coverageSignature — V8 snapshot -> fingerprint set
- coverageDelta — set difference
- noveltyMultiplier — log-bounded score-to-weight
- summarizeCoverage — report-shaped state summary
- targetKey — (url, selector) map key
- type CoverageReport / CoverageScriptResult

chaosbringer/src/coverage.ts becomes a thin re-export shim that
forwards to @mizchi/playwright-v8-coverage. crawler.ts and
chaosbringer's own index.ts continue to import from "./coverage.js"
unchanged — pure refactor, zero behavior change.

Workspace dep added: chaosbringer depends on
@mizchi/playwright-v8-coverage via "workspace:^". On publish, pnpm
rewrites this to the actual published version.

Verified locally:
- pnpm build (recursive, 3 packages) clean
- pnpm test (recursive) -> 487/487 pass
  - chaosbringer 457
  - playwright-v8-coverage 17
  - server-faults 13
- NUL gate clean for both packages
- TypeScript declaration files emitted

Note: not yet on npm. First publish needs manual mizchi action to
bootstrap OIDC for the new package name (same pattern as
@mizchi/server-faults).

After this PR, future chaosbringer releases that bump
@mizchi/playwright-v8-coverage will need the latter to be on npm
first. For the next chaosbringer release until that happens, pnpm
will resolve workspace:^ at publish time and pin to whatever the
local workspace version is — which is fine for source builds but
breaks consumers fetching from npm. Resolution: publish
playwright-v8-coverage 0.1.0 before the next chaosbringer release
that includes this change.
