+++ packages/playwright-v8-coverage/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "@mizchi/playwright-v8-coverage",
+  "version": "0.1.0",
+  "description": "V8 precise-coverage collector for Playwright (CDP Profiler.takePreciseCoverage) with novelty-scoring helpers. Extracted from chaosbringer.",
+  "license": "MIT",
+  "author": "mizchi",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mizchi/chaosbringer.git",
+    "directory": "packages/playwright-v8-coverage"
+  },
+  "homepage": "https://github.com/mizchi/chaosbringer/tree/main/packages/playwright-v8-coverage#readme",
+  "bugs": {
+    "url": "https://github.com/mizchi/chaosbringer/issues"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "engines": {
+    "node": ">=20"
+  },
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "test": "vitest run",
+    "lint:nul": "node ../chaosbringer/scripts/check-no-nul.mjs src"
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "keywords": [
+    "playwright",
+    "v8",
+    "coverage",
+    "precise-coverage",
+    "cdp",
+    "novelty"
+  ],
+  "peerDependencies": {
+    "playwright": "^1.40.0"
+  },
+  "peerDependenciesMeta": {
+    "playwright": {
+      "optional": true
+    }
+  },
+  "devDependencies": {
+    "playwright": "^1.52.0",
+    "typescript": "^5.8.3",
+    "vitest": "^3.0.7"
+  }
+}
