+++ packages/playwright-v8-coverage/src/index.ts
@@ -1,0 +1,10 @@
+export {
+  CoverageCollector,
+  coverageDelta,
+  coverageSignature,
+  noveltyMultiplier,
+  summarizeCoverage,
+  targetKey,
+  type CoverageReport,
+  type CoverageScriptResult,
+} from "./coverage.js";
